+++ examples/cra-ts-kitchen-sink/tsconfig.json
@@ -1,5 +1,7 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
+    "paths": { "*": ["types/*"] },
     "target": "es5",
     "module": "esnext",
     "lib" : ["esnext", "dom"],
