+++ examples/cra-ts-kitchen-sink/types/@storybook/react.d.ts
@@ -1,0 +1,8 @@
+export const addDecorator: any;
+export const addParameters: any;
+export const configure: any;
+export const forceReRender: any;
+export const getStorybook: any;
+export const raw: any;
+export const setAddon: any;
+export const storiesOf: any;
